+++ packages/stream_feed/test/token_helper_test.dart
@@ -35,11 +35,7 @@ void main() {
 
       expect(
         payloadJson,
-        {
-          'exp': isA<int>(),
-          // 'iat': isA<int>(),
-          'user_id': 'userId'
-        },
+        {'exp': isA<int>(), 'user_id': 'userId'},
       );
       expect(payloadJson['user_id'], 'userId');
     });
@@ -58,8 +54,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        //'iat': isA<int>(),
         'action': '*',
         'resource': 'feed',
         'feed_id': '*',
@@ -80,8 +74,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        //  'iat': isA<int>(),
         'action': '*',
         'resource': 'follower',
         'feed_id': '*',
@@ -103,8 +95,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        //  'iat': isA<int>(),
         'action': '*',
         'resource': 'reactions',
         'feed_id': '*',
@@ -126,8 +116,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        // 'iat': isA<int>(),
         'action': 'write',
         'resource': 'activities',
         'feed_id': '*',
@@ -148,8 +136,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        //  'iat': isA<int>(),
         'action': '*',
         'resource': 'users',
         'feed_id': '*',
@@ -171,8 +157,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        //  'iat': isA<int>(),
         'action': '*',
         'resource': 'collections',
         'feed_id': '*',
@@ -195,8 +179,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        // 'iat': isA<int>(),
         'action': 'read',
         'resource': 'url',
         'feed_id': '*',
@@ -218,8 +200,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        //   'iat': isA<int>(),
         'action': '*',
         'resource': 'feed_targets',
         'feed_id': '*',
@@ -240,8 +220,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        //  'iat': isA<int>(),
         'action': '*',
         'resource': 'analytics',
         'feed_id': '*',
@@ -264,8 +242,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        // 'iat': isA<int>(),
         'action': '*',
         'resource': 'personalization',
         'feed_id': '*',
@@ -288,8 +264,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        //   'iat': isA<int>(),
         'action': '*',
         'resource': 'redirect_and_track',
         'feed_id': '*',
@@ -311,8 +285,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        // 'iat': isA<int>(),
         'action': 'write',
         'resource': 'analytics',
         'feed_id': '*',
@@ -332,8 +304,6 @@ void main() {
       final payloadStr = b64urlEncRfc7515Decode(payload);
       final payloadJson = json.decode(payloadStr);
       expect(payloadJson, {
-        'exp': isA<int>(),
-        // 'iat': isA<int>(),
         'action': '*',
         'resource': 'files',
         'feed_id': '*',