+++ packages/stream_feed/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.5.2: 30/03/2022
+
+- fix(serverside llc): `issueJwtHS256` wasn't using the `expiresAt` field and remove the default "exp" value.
+- new: pagination suuport for flat feed activities. For example:
+```dart
+final paginated = await flatFeed.getPaginatedEnrichedActivities();
+final nextParams = parseNext(paginated.next!);
+// parse next page
+await flatFeed.getPaginatedEnrichedActivities(limit: nextParams.limit,filter: nextParams.idLT);
+```
+
 ## 0.5.1+1: 24/03/2022
 
 - fix: the `JsonConverter<DateTime,String>` implemented in 0.4.0+1 that was supposed to handle utc dates parsing wasn't working properly. Now that it is actually fixed you can convert dates in the user's local timezone.